Renamed blob_detection.cpp to blob_detection.hpp
