+++ 4_cv_basics/8_blob_detection/src/blob_detection.hpp
